pyproject.toml
[project]
name = "obsidian-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.4"
dependencies = [
    "fastapi>=0.120.3",
    "fastapi-mcp>=0.4.0",
    "httpx>=0.28.1",
    "isort>=7.0.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.2.1",
    "python-frontmatter>=1.1.0",
    "uvicorn>=0.38.0",
]
